requirements.txt

Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==10.0.0
gunicorn==21.2.0